Modularize ChannelStatus
