+++ src/rubintv/templates/admin.jinja
@@ -62,6 +62,6 @@ RubinTV Admin
 {% endblock content %}
 {% block footer_scripts %}
 <script src="//ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
-<script type=text/javascript src="{{ url('static', filename='js/modules/heartbeat.js') }}"></script>
+<script type=module src="{{ url('static', filename='js/admin-status.js') }}"></script>
 <script type=text/javascript src="{{ url('static', filename='js/modules/historicalReset.js') }}"></script>
 {% endblock footer_scripts %}
